@@ -12,8 +12,8 @@
     "clsx": "^1.1.1",
     "moment": "^2.29.3",
     "moment-jalaali": "^0.9.6",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
