@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import "./index.css";
 import App from "./App";
 import { ThemeProvider } from "@material-ui/core";
@@ -7,9 +7,10 @@ import { theme } from "./theme/default";
 import moment from "moment-jalaali";
 moment.loadPersian({ dialect: "persian-modern" });
 
-ReactDOM.render(
-  <ThemeProvider theme={theme}>
-    <App />
-  </ThemeProvider>,
-  document.getElementById("root")
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(  <ThemeProvider theme={theme}>
+  <App />
+</ThemeProvider>,
+document.getElementById("root")
 );
+
